requirements.txt
bcrypt==4.1.2
blinker==1.7.0
Cerberus==1.3.5
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pillow==10.2.0
pip==22.0.2
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
setuptools==59.6.0
Werkzeug==3.0.1
WTForms==3.1.2
Flask-Cors==4.0.0
